Updated README
I added a bunch of headings (most of which we will probably not use until later) in order to set up a general layout that corresponded roughly with what I read online. Also, I added a short description of the project and motivation. The motivation part needs to be changed.